[ICU-3500] Fix some compiler warnings X-Commit-URL: https://ssl.icu-project.org/trac/changeset/15406 Former-commit-id: a0dfdca1b3bb6186ad5f6e95dab67a38c03b657c

+++ icu4c/source/test/cintltst/ncnvfbts.c
@@ -115,11 +115,11 @@ void addTestConverterFallBack(TestNode** root)
 
 static void setNuConvTestName(const char *codepage, const char *direction)
 {
-  sprintf(gNuConvTestName, "[Testing %s %s Unicode, InputBufSiz=%d, OutputBufSiz=%d]",
-      codepage,
-      direction,
-      gInBufferSize,
-      gOutBufferSize);
+    sprintf(gNuConvTestName, "[Testing %s %s Unicode, InputBufSiz=%d, OutputBufSiz=%d]",
+        codepage,
+        direction,
+        (int)gInBufferSize,
+        (int)gOutBufferSize);
 }
